generalize whitelist check
